surfaceflinger: fix vsyncTime error [1/1]
PD#SWPL-114943

Problem:
drop frame when play 4k video, drop one frame every two frames sometimes

Solution:
get vsyncTime error, and calc display pts error,
two frames have same display pts

Verify:
T5M

+++ services/surfaceflinger/SurfaceFlinger.cpp
@@ -8571,8 +8571,8 @@ binder::Status SurfaceComposerAIDL::getDisplayStats(const sp<IBinder>& display,
     DisplayStatInfo statInfo;
     status_t status = mFlinger->getDisplayStats(display, &statInfo);
     if (status == NO_ERROR) {
-        outStatInfo->vsyncTime = static_cast<long>(statInfo.vsyncTime);
-        outStatInfo->vsyncPeriod = static_cast<long>(statInfo.vsyncPeriod);
+        outStatInfo->vsyncTime = static_cast<long long>(statInfo.vsyncTime);
+        outStatInfo->vsyncPeriod = static_cast<long long>(statInfo.vsyncPeriod);
     }
     return binderStatusFromStatusT(status);
 }
